javadocs corrected, messageSourse rebased to rest config
