+++ core/src/main/java/com/epam/esm/config/CoreSpringConfig.java
@@ -1,9 +1,7 @@
 package com.epam.esm.config;
 
-import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -18,13 +16,5 @@ public class CoreSpringConfig {
 		return builder;
 	}
 
-	@Bean
-	public MessageSource getMessageSource() {
-		ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-		messageSource.setBasename("classpath:messages");
-		messageSource.setDefaultEncoding("UTF-8");
-		messageSource.setCacheMillis(500);
-		return messageSource;
-	}
 
 }
