+++ dal/src/main/java/com/epam/esm/dal/config/DalSpringConfig.java
@@ -4,7 +4,6 @@ import javax.naming.NamingException;
 
 import org.apache.tomcat.dbcp.dbcp2.BasicDataSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -23,9 +22,6 @@ public class DalSpringConfig {
 
 	@Autowired
 	private Environment env;
-	
-	@Autowired
-	MessageSource messageSource;
 
 	@Bean
 	public BasicDataSource getDataSource() {
