+++ rest/src/main/java/com/epam/esm/rest/config/RestSpringConfig.java
@@ -1,15 +1,25 @@
 package com.epam.esm.rest.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.util.Locale;
+
 import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 @Configuration
 @EnableWebMvc
 public class RestSpringConfig {
 	
-	@Autowired
-	MessageSource messageSource;
+	@Bean
+	public MessageSource getMessageSource() {
+		ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
+		messageSource.setBasename("classpath:messages");
+		messageSource.setDefaultEncoding("UTF-8");
+		messageSource.setDefaultLocale(new Locale("en"));
+		messageSource.setCacheMillis(500);
+		return messageSource;
+	}
 
 }
