+++ service/src/main/java/com/epam/esm/service/config/ServiceSpringConfig.java
@@ -1,8 +1,6 @@
 package com.epam.esm.service.config;
 
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -10,9 +8,6 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @ComponentScan("com.epam.esm")
 public class ServiceSpringConfig {
-	
-	@Autowired
-	MessageSource messageSource;
 
 	@Bean
 	public ModelMapper getModelMapper() {
